correction nomage des packets
