+++ GalactikDiaballik.java
@@ -1,3 +1,4 @@
+import controller.listeners.MouseAction;
 import model.Stadium;
 import view.HoloTV;
 
@@ -5,7 +6,7 @@ public class GalactikDiaballik {
 	public static void main(String[] args) {
 		Stadium stadium = new Stadium();
 		HoloTV holoTV = new HoloTV(stadium);
-		Listeners.MouseAction mouseAction = new Listeners.MouseAction(holoTV, stadium);
+		MouseAction mouseAction = new MouseAction(holoTV, stadium);
 
 		holoTV.run();
 		holoTV.addArkadiaNewsMouseListener(mouseAction);
