+++ view/HoloTV.java
@@ -1,6 +1,8 @@
 package view;
 
 import javax.swing.JFrame;
+
+import controller.listeners.MouseAction;
 import model.Stadium;
 
 public class HoloTV implements Runnable {
@@ -32,7 +34,7 @@ public class HoloTV implements Runnable {
         frame.setVisible(true);
 	}
 	
-	public void addArkadiaNewsMouseListener(Listeners.MouseAction mouseAction) {
+	public void addArkadiaNewsMouseListener(MouseAction mouseAction) {
 		this.gamePanel.getArkadiaNews().addMouseListener(mouseAction);
 	}
